app/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
billiard==3.6.4.0
bitarray==1.2.2
boto3==1.21.16
botocore==1.24.16
CacheControl==0.12.11
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
cytoolz==0.12.0
daphne==3.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.6
django-allauth==0.51.0
django-appconf==1.0.5
django-cors-headers==3.13.0
django-cryptography==1.1
django-filter==22.1
django-location-field==2.1.0
django-heroku==0.3.1
django-storages==1.12.3
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
fcm-django==1.0.12
firebase-admin==5.2.0
frozenlist==1.3.0
geographiclib==1.52
geopy==2.2.0
google-api-core==2.8.2
google-api-python-client==2.53.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
hexbytes==0.2.2
httplib2==0.20.4
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==4.7.2
kombu==5.2.4
lru-dict==1.1.8
msgpack==1.0.4
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
oauthlib==3.2.0
parsimonious==0.8.1
Pillow==9.2.0
prompt-toolkit==3.0.30
proto-plus==1.20.6
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-decouple==3.6
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rlp==2.0.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
toolz==0.12.0
uritemplate==4.1.1
urllib3==1.26.10
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.30.0
websockets==9.1
yarl==1.7.2
numpy~=1.23.2
TurkishStemmer~=1.3
pandas~=1.5.0
scikit-learn~=1.1.2